Refactor shell server registration types to separate registration project
