+++ SharpShell/SharpShell.Registration/Diagnostics/ExplorerConfigurationManager.cs
