+++ SharpShell/SharpShell.Registration/Diagnostics/ExplorerManager.cs
