+++ SharpShell/SharpShell.Registration/ServerRegistration/ClassRegistration.cs
