+++ SharpShell/SharpShell.Registration/ServerRegistration/ServerRegistationType.cs
