+++ SharpShell/SharpShell.Registration/ServerRegistration/ShellExtensionType.cs
