+++ SharpShell/SharpShell.Tests/SharpShell.Tests.csproj
@@ -23,6 +23,7 @@
     </Content>
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\SharpShell.Registration\SharpShell.Registration.csproj" />
     <ProjectReference Include="..\SharpShell\SharpShell.csproj" />
   </ItemGroup>
   <ItemGroup>
