+++ SharpShell/SharpShell/Extensions/EnumExtensions.cs
