+++ SharpShell/SharpShell/Registry/IRegistry.cs
