+++ SharpShell/SharpShell/Registry/IRegistryKey.cs
