+++ SharpShell/SharpShell/Registry/InMemoryRegistry.cs
