+++ SharpShell/SharpShell/Registry/InMemoryRegistryKey.cs
