+++ SharpShell/SharpShell/Registry/WindowsRegistry.cs
