+++ SharpShell/SharpShell/Registry/WindowsRegistryKey.cs
