+++ SharpShell/SharpShell/ServiceRegistry/ServiceRegistry.cs
