+++ SharpShell/Tools/ServerManager/ServerManager.csproj
@@ -161,6 +161,7 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\..\SharpShell.Registration\SharpShell.Registration.csproj" />
     <ProjectReference Include="..\..\SharpShell\SharpShell.csproj" />
   </ItemGroup>
   <ItemGroup>
